+++ src/serialize.rs
@@ -1,6 +1,6 @@
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, BTreeSet};
 
-use crate::{Format, Key, KeyName, Kind, RawValue, ValueName};
+use crate::{Format, Key, KeyName, Kind, RawValue, ValueName, WineGlobalOption, WineKeyOption};
 
 fn quoted(raw: &str) -> String {
     format!("\"{raw}\"")
@@ -14,25 +14,67 @@ pub fn format(format: Format) -> &'static str {
     match format {
         Format::Regedit5 => Format::REGEDIT5,
         Format::Regedit4 => Format::REGEDIT4,
+        Format::Wine2 => Format::WINE2,
     }
+}
+
+pub fn key_name(name: &KeyName, addendum: Option<&String>) -> String {
+    let name = &name.0;
+
+    match addendum {
+        Some(addendum) => format!("[{name}] {addendum}"),
+        None => format!("[{name}]"),
+    }
+}
+
+pub fn key_name_deleted(name: &KeyName) -> String {
+    format!("[-{}]", name.0)
 }
 
 pub fn key(name: &KeyName, key: &Key, format: Format) -> String {
     match key {
-        Key::Delete => format!("[-{}]", name.0),
-        Key::Add { values: xs } => {
-            let mut lines = vec![format!("[{}]", name.0)];
+        Key::Delete => key_name_deleted(name),
+        Key::Add {
+            values,
+            addendum,
+            wine_options,
+        } => {
+            let mut lines = vec![key_name(name, addendum.as_ref())];
 
-            let xs: BTreeMap<_, _> = xs.clone().into_iter().map(|(n, v)| (n, v.into_raw(format))).collect();
-            lines.extend(values(&xs));
+            if format.is_wine() {
+                lines.extend(wine_key_options(wine_options));
+            }
+
+            let values: BTreeMap<_, _> = values
+                .clone()
+                .into_iter()
+                .map(|(n, v)| (n, v.into_raw(format)))
+                .collect();
+            lines.extend(self::values(&values));
 
             lines.join("\n")
         }
-        Key::Replace { values: xs } => {
-            let mut lines = vec![format!("[-{}]\n", name.0), format!("[{}]", name.0)];
+        Key::Replace {
+            values,
+            addendum,
+            wine_options,
+        } => {
+            let mut lines = vec![
+                key_name_deleted(name),
+                "".to_string(),
+                key_name(name, addendum.as_ref()),
+            ];
 
-            let xs: BTreeMap<_, _> = xs.clone().into_iter().map(|(n, v)| (n, v.into_raw(format))).collect();
-            lines.extend(values(&xs));
+            if format.is_wine() {
+                lines.extend(wine_key_options(wine_options));
+            }
+
+            let values: BTreeMap<_, _> = values
+                .clone()
+                .into_iter()
+                .map(|(n, v)| (n, v.into_raw(format)))
+                .collect();
+            lines.extend(self::values(&values));
 
             lines.join("\n")
         }
@@ -63,6 +105,10 @@ pub fn value(value: &RawValue, offset: usize) -> String {
         RawValue::Delete => "-".to_string(),
         RawValue::Sz(data) => quoted(&escape(data)),
         RawValue::Dword(data) => format!("dword:{:0>8x}", data),
+        RawValue::Str { kind, data } => match kind {
+            Kind::Sz => format!("str:{}", quoted(&escape(data))),
+            kind => format!("str({:x}):{}", u8::from(*kind), quoted(&escape(data))),
+        },
         RawValue::Hex { kind, bytes } => {
             let mut out = String::new();
 
@@ -92,6 +138,42 @@ pub fn value(value: &RawValue, offset: usize) -> String {
     }
 }
 
+pub fn wine_global_options(options: &BTreeSet<WineGlobalOption>) -> Vec<String> {
+    let mut lines = vec![];
+
+    for option in options {
+        lines.push(wine_global_option(option));
+    }
+
+    lines
+}
+
+pub fn wine_global_option(option: &WineGlobalOption) -> String {
+    match option {
+        WineGlobalOption::Arch(arch) => format!("#arch={arch}"),
+        WineGlobalOption::Other(other) => format!("#{other}"),
+    }
+}
+
+pub fn wine_key_options(options: &BTreeSet<WineKeyOption>) -> Vec<String> {
+    let mut lines = vec![];
+
+    for option in options {
+        lines.push(wine_key_option(option));
+    }
+
+    lines
+}
+
+pub fn wine_key_option(option: &WineKeyOption) -> String {
+    match option {
+        WineKeyOption::Class(class) => format!("#class=\"{class}\""),
+        WineKeyOption::Time(time) => format!("#time={time}"),
+        WineKeyOption::Link => "#link".to_string(),
+        WineKeyOption::Other(other) => format!("#{other}"),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -100,7 +182,8 @@ mod tests {
 
     #[test_case("[foo]", "foo", Key::new() ; "add")]
     #[test_case("[-foo]", "foo", Key::Delete ; "delete")]
-    fn invalid_keys(raw: &str, name: &str, parsed: Key) {
+    #[test_case("[foo] bar", "foo", Key::new().with_addendum("bar".to_string()) ; "addendum")]
+    fn valid_keys(raw: &str, name: &str, parsed: Key) {
         assert_eq!(
             raw.to_string(),
             key(&KeyName(name.to_string()), &parsed, Format::Regedit5)
@@ -117,6 +200,9 @@ mod tests {
 
     #[test_case("\"foo\"", RawValue::Sz("foo".to_string()) ; "sz simple")]
     #[test_case(r#""sp\\ec\"ial""#, RawValue::Sz(r#"sp\ec"ial"#.to_string()) ; "sz escaped characters")]
+    #[test_case("str:\"foo\"", RawValue::Str { data: "foo".to_string(), kind: Kind::Sz } ; "str sz")]
+    #[test_case("str(2):\"foo\"", RawValue::Str { data: "foo".to_string(), kind: Kind::ExpandSz } ; "str expand")]
+    #[test_case("str(7):\"foo\"", RawValue::Str { data: "foo".to_string(), kind: Kind::MultiSz } ; "str multi")]
     #[test_case("dword:00000000", RawValue::Dword(0) ; "dword 0")]
     #[test_case("dword:000000ff", RawValue::Dword(255) ; "dword 255")]
     #[test_case("-", RawValue::Delete ; "delete")]
@@ -129,5 +215,19 @@ mod tests {
     #[test_case("hex(b):ff,00,00,00,00,00,00,00", RawValue::Hex { kind: Kind::Qword, bytes: vec![255, 0, 0, 0, 0, 0, 0, 0] } ; "hex qword 255")]
     fn valid_values(raw: &str, parsed: RawValue) {
         assert_eq!(raw, value(&parsed, 0));
+    }
+
+    #[test_case("#arch=win32", WineGlobalOption::Arch("win32".to_string()) ; "arch")]
+    #[test_case("#foo", WineGlobalOption::Other("foo".to_string()) ; "other")]
+    fn valid_wine_global_options(raw: &str, parsed: WineGlobalOption) {
+        assert_eq!(raw, wine_global_option(&parsed));
+    }
+
+    #[test_case("#time=100", WineKeyOption::Time(100) ; "time")]
+    #[test_case("#class=\"foo\"", WineKeyOption::Class("foo".to_string()) ; "class")]
+    #[test_case("#link", WineKeyOption::Link ; "link")]
+    #[test_case("#foo", WineKeyOption::Other("foo".to_string()) ; "other")]
+    fn valid_wine_key_options(raw: &str, parsed: WineKeyOption) {
+        assert_eq!(raw, wine_key_option(&parsed));
     }
 }